+++ system_management/app/Database/Migrations/2024-07-23-023622_AddIdToSoTTable.php
@@ -18,9 +18,17 @@ class AddIdToSoTTable extends Migration
 
     public function down()
     {
-        // Remove id field from banks table
-        $this->db->query('ALTER TABLE so_t DROP COLUMN id');
+        $columnExists = $this->db->query("SELECT COUNT(*) AS count 
+                                          FROM information_schema.columns 
+                                          WHERE table_name = 'so_t' 
+                                          AND column_name = 'id'")
+                                 ->getRow()
+                                 ->count;
 
+        if ($columnExists > 0) {
+            // Drop the 'id' column if it exists
+            $this->db->query('ALTER TABLE so_t DROP COLUMN id');
+        }
         // Restore primary key of banks table
         $this->db->query('ALTER TABLE so_t ADD PRIMARY KEY(compcode, cidentity)');
